3.5.3  Added an SCA reset enable mask, and updated SCA GPIO defaults for VFAT3 reset
Added an SCA reset enable mask register which defines which SCAs will
get reset uppon a receipt of a module reset command.

Also updated the default GPIO direction and output value to drive the
VFAT3 resets in OHv3c. The reset is lifted by default after SCA reset,
but it is pulsed high when a hard reset command is received.